+++ Editor/IAPProductButtonEditor.cs
@@ -167,3 +167,4 @@ namespace Wagenheimer.IAPHelper.Editor
         }
     }
 }
+